Continues unifying ReaderIterator and Source.Reader
* Flattens the ReaderIterator hierarchy a bit:
  removes Abstract{Bounded,}ReaderIterator, instead merges them all
  into LegacyReaderIterator directly.
* Converts InMemoryReaderIterator, PubsubReaderIterator,
  UngroupedWindmillReaderIterator, ConcatIterator, and a couple of
  anonymous classes to the new interface.
  File-based and shuffle readers remain.
* Extracts a common base class from Pubsub and UngroupedWindmill readers.
* Converts most call sites that could use Source.Reader instead of
  ReaderIterator to use either Source.Reader, or the concrete class
  of the reader, to ease future conversion.

----Release Notes----
[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=112728940

+++ sdk/src/main/java/com/google/cloud/dataflow/sdk/runners/worker/InMemoryReaderFactory.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.dataflow.sdk.runners.worker;
 
-import static com.google.cloud.dataflow.sdk.util.Structs.getLong;
+import static com.google.cloud.dataflow.sdk.util.Structs.getInt;
 import static com.google.cloud.dataflow.sdk.util.Structs.getStrings;
 
 import com.google.cloud.dataflow.sdk.coders.Coder;
@@ -51,7 +51,8 @@ public class InMemoryReaderFactory implements ReaderFactory {
   <T> InMemoryReader<T> create(CloudObject spec, Coder<T> coder) throws Exception {
     return new InMemoryReader<>(
         getStrings(spec, PropertyNames.ELEMENTS, Collections.<String>emptyList()),
-        getLong(spec, PropertyNames.START_INDEX, null),
-        getLong(spec, PropertyNames.END_INDEX, null), coder);
+        getInt(spec, PropertyNames.START_INDEX, null),
+        getInt(spec, PropertyNames.END_INDEX, null),
+        coder);
   }
 }
